requirements.txt
discord.py==1.7.3
googletrans==4.0.0-rc1
google-cloud-texttospeech==2.7.0
